python/requirements.txt
todoist-python
python-dotenv
requests
